[fix]:add linux run example

+++ run_example.sh
@@ -65,16 +65,23 @@ cp model/mobilenet_iter_73000.dlc ${DEMO_DIR}
 cp data/VOC_raw/1.raw ${DEMO_DIR}
 cp data/VOC_resize/1.jpg ${DEMO_DIR}
 cp third_party/snpe/1.43/lib/${TARGET}/* ${DEMO_DIR}
+cp third_party/opencv_2.4.13.4/${TARGET}/lib/* ${DEMO_DIR}
+cp third_party/opencv_2.4.13.4/${TARGET}/3rdparty/* ${DEMO_DIR}
 
-bin_path="/data/local/tmp/test_snpe_demo/"
-adb shell "rm -r ${bin_path}"
-adb push ${DEMO_DIR} /data/local/tmp/
-adb shell "chmod +x ${bin_path}/test-mobilenet-ssd"
-#adb shell "cd ${bin_path} \
-#       && export LD_LIBRARY_PATH=${bin_path}:${LD_LIBRARY_PATH} \
-#       && ./test-mobilenet-ssd mobilenet_iter_73000_int8.dlc 1.raw"
-adb shell "cd ${bin_path} \
-       && export LD_LIBRARY_PATH=${bin_path}:${LD_LIBRARY_PATH} \
-       && ./test-mobilenet-ssd 2 mobilenet_iter_73000_int8.dlc 1.jpg"
+if [ "${TARGET}" = "linux-x86_64" ]; then
+  cd ${DEMO_DIR}
+  export LD_LIBRARY_PATH=${PWD}:${LD_LIBRARY_PATH}
+  ./test-mobilenet-ssd 0 mobilenet_iter_73000_int8.dlc 1.jpg
+else
+  bin_path="/data/local/tmp/test_snpe_demo/"
+  adb shell "rm -r ${bin_path}"
+  adb push ${DEMO_DIR} /data/local/tmp/
+  adb shell "chmod +x ${bin_path}/test-mobilenet-ssd"
+  adb shell "cd ${bin_path} \
+         && export LD_LIBRARY_PATH=${bin_path}:${LD_LIBRARY_PATH} \
+         && ./test-mobilenet-ssd 2 mobilenet_iter_73000_int8.dlc 1.jpg"
+  adb pull $bin_path"result.jpg" ./
+fi
 
-adb pull $bin_path"result.jpg" ./
+
+
